intelligent-service domain class modified

+++ service-provider-profile-service/src/main/java/com/stackroute/serviceproviderprofileservice/domain/ServiceProvider.java
@@ -8,6 +8,7 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import javax.validation.constraints.Email;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -25,6 +26,6 @@ public class ServiceProvider {
      String role;
      ArrayList<SkillsAndExperience> skillsAndExperiences;
      String about;
-     Double chargePerHour;
+     BigDecimal chargePerHour;
 
 }
